[Rename] : main/excption 합침 (#22)
