+++ src/main/java/com/potato/balbambalbam/main/MyConstant.java
@@ -1,4 +1,4 @@
-package com.potato.balbambalbam;
+package com.potato.balbambalbam.main;
 
 public final class MyConstant {
     public static final String AI_URL = "http://211.110.165.175:5000";
