+++ src/main/java/com/potato/balbambalbam/main/cardFeedback/service/AiCardFeedbackService.java
@@ -1,7 +1,7 @@
 package com.potato.balbambalbam.main.cardFeedback.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.potato.balbambalbam.MyConstant;
+import com.potato.balbambalbam.main.MyConstant;
 import com.potato.balbambalbam.main.cardFeedback.dto.AiFeedbackRequestDto;
 import com.potato.balbambalbam.main.cardFeedback.dto.AiFeedbackResponseDto;
 import lombok.extern.slf4j.Slf4j;
