+++ src/main/java/com/potato/balbambalbam/main/cardInfo/service/CardInfoService.java
@@ -8,8 +8,8 @@ import com.potato.balbambalbam.data.repository.UserRepository;
 import com.potato.balbambalbam.main.cardInfo.dto.AiTtsRequestDto;
 import com.potato.balbambalbam.main.cardInfo.dto.AiTtsResponseDto;
 import com.potato.balbambalbam.main.cardInfo.dto.CardInfoResponseDto;
-import com.potato.balbambalbam.main.cardInfo.exception.UserNotFoundException;
-import com.potato.balbambalbam.main.cardList.exception.CardNotFoundException;
+import com.potato.balbambalbam.main.exception.UserNotFoundException;
+import com.potato.balbambalbam.main.exception.CardNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
