+++ src/main/java/com/potato/balbambalbam/main/exception/AiConnectionException.java
@@ -1,4 +1,4 @@
-package com.potato.balbambalbam.main.cardInfo.exception;
+package com.potato.balbambalbam.main.exception;
 
 public class AiConnectionException extends RuntimeException {
     public AiConnectionException(String message) {
