+++ src/main/java/com/potato/balbambalbam/main/exception/CardNotFoundException.java
@@ -1,4 +1,4 @@
-package com.potato.balbambalbam.main.cardList.exception;
+package com.potato.balbambalbam.main.exception;
 
 public class CardNotFoundException extends IllegalArgumentException{
     public CardNotFoundException(String s) {
