+++ src/main/java/com/potato/balbambalbam/main/exception/CategoryNotFoundException.java
@@ -1,4 +1,4 @@
-package com.potato.balbambalbam.main.cardList.exception;
+package com.potato.balbambalbam.main.exception;
 
 public class CategoryNotFoundException extends IllegalArgumentException{
     public CategoryNotFoundException(String message) {
