+++ src/main/java/com/potato/balbambalbam/main/exception/ExceptionDto.java
@@ -1,4 +1,4 @@
-package com.potato.balbambalbam.main;
+package com.potato.balbambalbam.main.exception;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
