+++ src/main/java/com/potato/balbambalbam/main/exception/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package com.potato.balbambalbam.main.cardInfo.exception;
+package com.potato.balbambalbam.main.exception;
 
 public class UserNotFoundException extends RuntimeException{
     public UserNotFoundException(String message) {
